requirements.txt
pandas
polars
numpy
matplotlib
scikit-learn
jinja2
weasyprint
python-pptx
openai